+++ tests/models/classes/paymentMethod/OneyPaymentMethod/getErrorLabelTest.php
@@ -7,8 +7,6 @@ namespace PayPlug\tests\models\classes\paymentMethod\OneyPaymentMethod;
  * @group class
  * @group payment_method_class
  * @group oney_payment_method_class
- *
- * @runTestsInSeparateProcesses
  */
 class getErrorLabelTest extends BaseOneyPaymentMethod
 {