+++ changelog.md
@@ -4,6 +4,7 @@
 - Refactor :
   - [PRE-3040](https://gitlab.com/dalenys/public/ecommerce/prestashop_v4_17_8/-/merge_requests/2723): Add log in logout action
   - [PRE-3064](https://gitlab.com/dalenys/public/ecommerce/prestashop_v4_17_8/-/merge_requests/2721): Update JWT experacy date and update payplug-php library to 4.0.0
+  - [PRE-2206](https://gitlab.com/dalenys/public/ecommerce/prestashop_v4_17_8/-/merge_requests/2728): Refactor Oney unit tests
 
 ## Version 4.20.1
 - Bugfix :